// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        TipoHabitacionPublico: {
            id: components["schemas"]["Id"];
            /** @example Doble Superior */
            nombre: string;
            descripcion: string;
            /**
             * @description Capacidad máxima de huéspedes.
             * @example 4
             */
            capacidad: number;
            precioBaseNoche: components["schemas"]["Monto"];
            /** @description Puede estar vacía (la web usa una imagen genérica). */
            fotos: components["schemas"]["UrlFoto"][];
        };
// ...
        ReservaWebCreada: {
            codigo: components["schemas"]["CodigoReserva"];
            estado: components["schemas"]["EstadoReserva"];
            tipoHabitacion: components["schemas"]["TipoHabitacionReferencia"];
            entrada: components["schemas"]["FechaEstadia"];
            salida: components["schemas"]["FechaEstadia"];
            noches: number;
            numeroHuespedes: number;
            total: components["schemas"]["Monto"];
            /** @description Momento en que la reserva se cancela sola si no se paga (30 minutos). */
            pagoVenceEn: components["schemas"]["FechaHora"];
        };
// ...
        ReservaAppDetalle: {
            codigo: components["schemas"]["CodigoReserva"];
            entrada: components["schemas"]["FechaEstadia"];
            salida: components["schemas"]["FechaEstadia"];
            estado: components["schemas"]["EstadoReserva"];
            /** @constant */
            horaCheckOut: "12:00";
            tipoHabitacion: components["schemas"]["TipoHabitacionReferencia"];
            numeroHuespedes: number;
            habitacion: components["schemas"]["HabitacionReferencia"] | null;
        };
// ...
    };
// ...
}